// apps/mobile/gen/api/schemas/updateTodo.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Techoo API
 * API for the Techoo task management application with OpenAPI documentation
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateTodo {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  starts_at?: string | null;
  /** @nullable */
  ends_at?: string | null;
  /**
   * @minimum 0
   * @maximum 1
   */
  is_all_day?: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  done?: number;
}